base64 decode arg casts
```
src/knownhost.c:173:..: warning C4267: 'function': conversion from 'size_t' to 'unsigned int', possible loss of data
src/knownhost.c:173:42: warning: conversion from 'size_t' {aka 'long long unsigned int'} to 'unsigned int' may change value [-Wconversion]
src/knownhost.c:173:42: warning: conversion from 'size_t' {aka 'long unsigned int'} to 'unsigned int' may change value [-Wconversion]
src/knownhost.c:173:42: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long long') to 'unsigned int' [-Wshorten-64-to-32]
src/knownhost.c:173:42: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long') to 'unsigned int' [-Wshorten-64-to-32]
src/knownhost.c:180:..: warning C4267: 'function': conversion from 'size_t' to 'unsigned int', possible loss of data
src/knownhost.c:180:42: warning: conversion from 'size_t' {aka 'long long unsigned int'} to 'unsigned int' may change value [-Wconversion]
src/knownhost.c:180:42: warning: conversion from 'size_t' {aka 'long unsigned int'} to 'unsigned int' may change value [-Wconversion]
src/knownhost.c:180:42: warning: implicit conversion loses integer precision: 'unsigned long long' to 'unsigned int' [-Wshorten-64-to-32]
src/knownhost.c:180:42: warning: implicit conversion loses integer precision: 'unsigned long' to 'unsigned int' [-Wshorten-64-to-32]
src/pem.c:394:..: warning C4267: 'function': conversion from 'size_t' to 'unsigned int', possible loss of data
src/pem.c:394:39: warning: conversion from 'size_t' {aka 'long long unsigned int'} to 'unsigned int' may change value [-Wconversion]
src/pem.c:394:39: warning: conversion from 'size_t' {aka 'long unsigned int'} to 'unsigned int' may change value [-Wconversion]
src/pem.c:394:39: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long long') to 'unsigned int' [-Wshorten-64-to-32]
src/pem.c:394:39: warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long') to 'unsigned int' [-Wshorten-64-to-32]
src/userauth.c:614:..: warning C4244: 'function': conversion from '__int64' to 'unsigned int', possible loss of data
src/userauth.c:614:49: warning: conversion from 'long int' to 'unsigned int' may change value [-Wconversion]
src/userauth.c:614:49: warning: conversion from 'long long int' to 'unsigned int' may change value [-Wconversion]
src/userauth.c:614:49: warning: implicit conversion loses integer precision: 'long long' to 'unsigned int' [-Wshorten-64-to-32]
src/userauth.c:614:49: warning: implicit conversion loses integer precision: 'long' to 'unsigned int' [-Wshorten-64-to-32]
src/userauth.c:718:..: warning C4244: 'function': conversion from '__int64' to 'unsigned int', possible loss of data
src/userauth.c:718:49: warning: conversion from 'long int' to 'unsigned int' may change value [-Wconversion]
src/userauth.c:718:49: warning: conversion from 'long long int' to 'unsigned int' may change value [-Wconversion]
src/userauth.c:718:49: warning: implicit conversion loses integer precision: 'long long' to 'unsigned int' [-Wshorten-64-to-32]
src/userauth.c:718:49: warning: implicit conversion loses integer precision: 'long' to 'unsigned int' [-Wshorten-64-to-32]
```

+++ src/pem.c
@@ -391,7 +391,7 @@ _libssh2_openssh_pem_parse_data(LIBSSH2_SESSION * session,
 
     /* decode file */
     if(libssh2_base64_decode(session, (char **)&f, &f_len,
-                             b64data, b64datalen)) {
+                             b64data, (unsigned int)b64datalen)) {
        ret = -1;
        goto out;
     }
